+++ classes/model/CreditAccount.h
@@ -16,8 +16,8 @@ class CreditAccount: public Account
 {
 private:	
     Customer *_owner;
-    int _gettedMoney;
-    int _limit;
+    size_t _gettedMoney;
+    size_t _limit;
     Date _date;
 
     vector<string> _history;
@@ -28,35 +28,35 @@ private:
 	CreditAccount(const CreditAccount&);
 	CreditAccount& operator=(const CreditAccount&);
 public:
-    CreditAccount(Customer&, int, string, string);
+    CreditAccount(Customer&, size_t, string, string);
 	~CreditAccount();
 
 
     /*
-     * getMoney(int quantity)
+     * getMoney(size_t quantity)
      * quantity - amount of money that will be getted
      * return:
      *          true - if transaction was done
      *          false - if transaction was not done
      */
-    bool getMoney(int quantity);
+    bool getMoney(size_t quantity);
 
     /*
-     * putMOney(int quantity)
+     * putMOney(size_t quantity)
      * quantity - amount of money that will be inserted
      * return:
      *          true - if transaction was done
      *          false - if transaction was not done
      */
-    bool putMoney(int quantity);
+    bool putMoney(size_t quantity);
 
     void addActionToHistory(const string& action);
 
     inline vector<string>& getHistory() { return _history; };
 
     // show balance
-    inline const int checkBalance() const { return (_limit - _gettedMoney); };
-    inline int gettedMoney() const { return _gettedMoney; };
+    inline const size_t checkBalance() const { return (_limit - _gettedMoney); };
+    inline size_t gettedMoney() const { return _gettedMoney; };
 
     inline const string& getLogin() { return _accountNumber; };
     inline const string& getPassword() { return _password; };
